// src/libs/api/model/permissions.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Permissions { 
    bulkUpload?: boolean;
    createAccount?: boolean;
    exportAccount?: boolean;
}